LMA-A.py
Write a function that computes the scalar product of two vectors of arbitrary but equal length.  
The function should use "assert" to ensure 
that the input values are tuples and 
that they are tuples of equal length (i.e. equal number of elements) 
that they contain only integers or floats.